Revert "Merge branch 'dev' of https://github.com/dataDeathKnight/yunsoo into dev"
This reverts commit ee4bba9303ca844dad5ee6cdf7525bc902c48d6c, reversing
changes made to 87e5f6995b471faa34b4b1cb788eff48fe7b63ee.

+++ api/src/main/java/com/yunsoo/api/domain/ProductDomain.java
@@ -48,7 +48,7 @@ public class ProductDomain {
         product.setCreatedDateTime(productObject.getCreatedDateTime().toString());
 
         //fill with ProductBase information.
-        int productBaseId = productObject.getProductBaseId();
+        long productBaseId = productObject.getProductBaseId();
         ProductBaseObject productBaseObject = dataAPIClient.get("productbase/{id}", ProductBaseObject.class, productBaseId);
         product.setProductBaseId(productBaseId);
         product.setBarcode(productBaseObject.getBarcode());
